Split TextUi class
Extracted out responsibility for formatting text for display (i.e. output to user) into a new Formatter class
TextUi class now only responsible for parsing user commands(user input)
